requirements.txt
openai==0.27.6
pandas==2.0.2
requests==2.30.0
typer==0.9.0
Pillow==9.5.0
httpx==0.24.1
pymongo==4.3.3